+++ modules/spring-web/src/main/kotlin/com/explyt/spring/web/SpringWebClasses.kt
@@ -35,6 +35,7 @@ object SpringWebClasses {
     const val RESPONSE_ENTITY = "org.springframework.http.ResponseEntity"
     const val FEIGN_CLIENT = "org.springframework.cloud.openfeign.FeignClient"
     const val JAVA_HTTP_CLIENT = "java.net.http.HttpClient"
+    const val HTTP_EXCHANGE = "org.springframework.web.service.annotation.HttpExchange"
 
     const val MOCK_MVC_REQUEST_BUILDERS = "org.springframework.test.web.servlet.request.MockMvcRequestBuilders"
     private const val WEB_CLIENT = "org.springframework.web.reactive.function.client.WebClient"